Warehousemicroservice/requirements.txt
Django==4.2.3
django-rest-framework==0.1.0
django-cors-headers==4.2.0
requests==2.31.0
redis==4.6.0
channels==4.0.0
celery==5.3.1
python-dotenv
psycopg2-binary
gunicorn==20.1.0
whitenoise==6.2.0
